Added espresso tests
Added espresso tests that still need to be worked on. However I have done extensive testing on the barcode stuff, I know it works, just needs to check tags (adding images via intents is a mystery to me). Added a text listener so that a barcode entered manually will produce the same result.